// src/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export enum AccountCategoryType {
  BUYER = "BUYER",
  SELLER = "SELLER",
}
// ...
export type ModelFloatInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export type ModelUserConditionInput = {
  name?: ModelStringInput | null,
  email?: ModelStringInput | null,
  phone_number?: ModelStringInput | null,
  rating?: ModelFloatInput | null,
  accountType?: ModelStringInput | null,
  lastOnlineAt?: ModelIntInput | null,
  lat?: ModelFloatInput | null,
  lng?: ModelFloatInput | null,
  ledgerBalance?: ModelFloatInput | null,
  address?: ModelStringInput | null,
  city?: ModelStringInput | null,
  state?: ModelStringInput | null,
  zipCode?: ModelStringInput | null,
  lga?: ModelStringInput | null,
  website?: ModelStringInput | null,
  incorporateDate?: ModelStringInput | null,
  rcNumber?: ModelStringInput | null,
  totalOrders?: ModelIntInput | null,
  level?: ModelStringInput | null,
  identification?: ModelStringInput | null,
  identificationNumber?: ModelStringInput | null,
  identityDocs?: ModelStringInput | null,
  keyProduct?: ModelStringInput | null,
  country?: ModelStringInput | null,
  inviteCode?: ModelStringInput | null,
  accountCategory?: ModelAccountCategoryTypeInput | null,
  title?: ModelStringInput | null,
  logo?: ModelStringInput | null,
  backgroundImage?: ModelStringInput | null,
  images?: ModelStringInput | null,
  businessType?: ModelStringInput | null,
  certifications?: ModelStringInput | null,
  certsDoc?: ModelStringInput | null,
  mainMarkets?: ModelStringInput | null,
  memberShipType?: ModelStringInput | null,
  sellerLevel?: ModelStringInput | null,
  estRevenue?: ModelFloatInput | null,
  totalStaff?: ModelStringInput | null,
  responseTime?: ModelStringInput | null,
  languages?: ModelStringInput | null,
  legalRep?: ModelStringInput | null,
  overview?: ModelStringInput | null,
  activeOrder?: ModelFloatInput | null,
  and?: Array< ModelUserConditionInput | null > | null,
  or?: Array< ModelUserConditionInput | null > | null,
  not?: ModelUserConditionInput | null,
};

export type ModelAccountCategoryTypeInput = {
  eq?: AccountCategoryType | null,
  ne?: AccountCategoryType | null,
};

export type UpdateUserInput = {
  id: string,
  name?: string | null,
  email?: string | null,
  phone_number?: string | null,
  rating?: number | null,
  accountType?: string | null,
  lastOnlineAt?: number | null,
  lat?: number | null,
  lng?: number | null,
  ledgerBalance?: number | null,
  address?: string | null,
  city?: string | null,
  state?: string | null,
  zipCode?: string | null,
  lga?: string | null,
  website?: string | null,
  incorporateDate?: string | null,
  rcNumber?: string | null,
  totalOrders?: number | null,
  level?: string | null,
  identification?: string | null,
  identificationNumber?: string | null,
  identityDocs?: Array< string | null > | null,
  keyProduct?: string | null,
  country?: string | null,
  inviteCode?: string | null,
  accountCategory?: AccountCategoryType | null,
  title?: string | null,
  logo?: string | null,
  backgroundImage?: string | null,
  images?: Array< string | null > | null,
  businessType?: string | null,
  certifications?: string | null,
  certsDoc?: Array< string | null > | null,
  mainMarkets?: Array< string | null > | null,
  memberShipType?: string | null,
  sellerLevel?: string | null,
  estRevenue?: number | null,
  totalStaff?: string | null,
  responseTime?: string | null,
  languages?: Array< string | null > | null,
  legalRep?: string | null,
  overview?: string | null,
  activeOrder?: number | null,
};

export type DeleteUserInput = {
  id: string,
};
// ...
export type UpdateUserMutationVariables = {
  input: UpdateUserInput,
  condition?: ModelUserConditionInput | null,
};
// ...
export type DeleteUserMutationVariables = {
  input: DeleteUserInput,
  condition?: ModelUserConditionInput | null,
};
// ...
export type GetUserQueryVariables = {
  id: string,
};